Extracted functions into Functions.cs
